fixed some bugs added validations for admin privleges

+++ frontend/src/components/AllUsersTable/AllUsersTable.tsx
@@ -7,7 +7,7 @@ import { IUser } from "../../interfaces";
 
 interface AllUsersTableProps {
   usersList: IUser[];
-  deleteUser: (email: string) => void;
+  deleteUser: (user: IUser) => void;
 }
 
 const AllUsersTable: React.FC<AllUsersTableProps> = ({
@@ -16,17 +16,18 @@ const AllUsersTable: React.FC<AllUsersTableProps> = ({
 }) => {
   const [showAll, setShowAll] = useState<boolean>(false);
   const [showModal, setShowModal] = useState<boolean>(false);
-  const [userToDelete, setUserToDelete] = useState<string>("");
+  const [userToDelete, setUserToDelete] = useState<IUser | null>(null);
   const { activeUser } = useUsersStore();
 
   const removeUserHandler = () => {
     setShowModal(false);
+    if (!userToDelete) return;
     deleteUser(userToDelete);
   };
 
-  const clickOnUserHandler = (email: string) => {
+  const clickOnUserHandler = (user: IUser) => {
     if (!activeUser || !activeUser.isAdmin) return;
-    setUserToDelete(email);
+    setUserToDelete(user);
     setShowModal(true);
   };
 
@@ -36,18 +37,12 @@ const AllUsersTable: React.FC<AllUsersTableProps> = ({
         {showAll
           ? usersList.map((user, index: number) => (
               <div key={index}>
-                <Profile
-                  onClick={() => clickOnUserHandler(user.email)}
-                  user={user}
-                />
+                <Profile onClick={() => clickOnUserHandler(user)} user={user} />
               </div>
             ))
           : usersList.slice(0, 5).map((user, index: number) => (
               <div key={index}>
-                <Profile
-                  onClick={() => clickOnUserHandler(user.email)}
-                  user={user}
-                />
+                <Profile onClick={() => clickOnUserHandler(user)} user={user} />
               </div>
             ))}
         {usersList.length > 5 && (
@@ -61,7 +56,7 @@ const AllUsersTable: React.FC<AllUsersTableProps> = ({
       </div>
       {showModal && (
         <Modal onClose={() => {}}>
-          <h2>{`Would you like to remove "${userToDelete}" from this project?`}</h2>
+          <h2>{`Would you like to remove "${userToDelete?.email}" from this project?`}</h2>
           <div className="flex justify-around">
             <Button onClick={removeUserHandler} type="submit">
               Yes
